requirements.txt
torch
numpy
pyyaml
tqdm
gcsfs
pandas
gemmi
rdkit-pypi
